Typing corrections & Refactoring: getSongsByAlbumId

+++ migrations/1689638005081_create-table-songs.js
@@ -6,10 +6,10 @@ exports.up = pgm => {
 	pgm.createTable("songs", {
 		id: { type:"VARCHAR(64)", primaryKey: true },
 		title: { type:"TEXT", notNull: true },
-		year: { type:"SMALLINT", notNull: true },
+		year: { type:"INTEGER", notNull: true },
 		genre: { type:"TEXT", notNull: true },
 		performer: { type:"TEXT", notNull: true },
-		duration: { type:"SMALLINT" },
+		duration: { type:"INTEGER" },
 		album_id: { type:"TEXT" }
 	})
 }
